# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Gestão e Monitoramento de Casos de Cólera",
        "description": "Documentação da API para Sistema de Gestão e Monitoramento de Casos de Cólera em Angola",
        "contact": {
            "name": "Suporte Técnico",
            "email": "admin@saudeangola.gov"
        },
        "license": {
            "name": "Licença MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "API Versão 1"
        }
    ],
    "paths": {
        "/auditoria": {
            "get": {
                "tags": [
                    "Auditoria"
                ],
                "summary": "Lista atividades de auditoria",
                "description": "Retorna uma lista paginada de atividades de auditoria com opções de filtragem",
                "operationId": "auditLogsList",
                "parameters": [
                    {
                        "name": "data_inicio",
                        "in": "query",
                        "description": "Data de início para filtrar (formato Y-m-d)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "data_fim",
                        "in": "query",
                        "description": "Data de fim para filtrar (formato Y-m-d)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "modelo",
                        "in": "query",
                        "description": "Nome do modelo para filtrar (ex: Paciente, Triagem)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "evento",
                        "in": "query",
                        "description": "Tipo de evento: created, updated ou deleted",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "created",
                                "updated",
                                "deleted"
                            ]
                        }
                    },
                    {
                        "name": "usuario_id",
                        "in": "query",
                        "description": "ID do usuário que realizou as ações",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Número de registros por página (10-100)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15,
                            "maximum": 100,
                            "minimum": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operação bem-sucedida",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Registros de auditoria obtidos com sucesso"
                                        },
                                        "data": {
                                            "properties": {
                                                "atividades": {
                                                    "properties": {
                                                        "current_page": {
                                                            "type": "integer"
                                                        },
                                                        "data": {
                                                            "type": "array",
                                                            "items": {
                                                                "type": "object"
                                                            }
                                                        },
                                                        "total": {
                                                            "type": "integer"
                                                        },
                                                        "per_page": {
                                                            "type": "integer"
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "filtros_aplicados": {
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Não autorizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Não autorizado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "429": {
                        "description": "Limite de requisições excedido"
                    }
                },
                "security": {
                    "bearerAuth": {}
                }
            }
        },
        "/auditoria/{id}": {
            "get": {
                "tags": [
                    "Auditoria"
                ],
                "summary": "Mostra detalhes de uma atividade",
                "description": "Retorna os detalhes de uma atividade de auditoria específica",
                "operationId": "auditLogsShow",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da atividade de auditoria",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operação bem-sucedida",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Detalhes do registro de auditoria obtidos com sucesso"
                                        },
                                        "data": {
                                            "properties": {
                                                "atividade": {
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Atividade não encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Registro de auditoria não encontrado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Não autorizado"
                    }
                },
                "security": {
                    "bearerAuth": {}
                }
            }
        },
        "/auditoria/resumo": {
            "get": {
                "tags": [
                    "Auditoria"
                ],
                "summary": "Resumo de atividades recentes",
                "description": "Retorna um resumo estatístico das atividades recentes",
                "operationId": "auditLogsSummary",
                "parameters": [
                    {
                        "name": "dias",
                        "in": "query",
                        "description": "Número de dias para analisar",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 30
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operação bem-sucedida",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Resumo de atividades obtido com sucesso"
                                        },
                                        "data": {
                                            "properties": {
                                                "periodo": {
                                                    "properties": {
                                                        "inicio": {
                                                            "type": "string",
                                                            "format": "date"
                                                        },
                                                        "fim": {
                                                            "type": "string",
                                                            "format": "date"
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "total_atividades": {
                                                    "type": "integer"
                                                },
                                                "por_modelo": {
                                                    "type": "object"
                                                },
                                                "por_evento": {
                                                    "type": "object"
                                                },
                                                "usuarios_mais_ativos": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Não autorizado"
                    }
                },
                "security": {
                    "bearerAuth": {}
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}